fix(init): fix system init

+++ system_init.sh
@@ -30,8 +30,10 @@ sudo apt install -y build-essential cmake python3 pip
 wget -O - https://go.dev/dl/go1.22.6.linux-amd64.tar.gz | sudo tar -zx -C /usr/local
 
 # zsh
+proxyon
 sudo apt install -y zsh \
 	&& sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+proxyoff
 
 # docker
 curl -fsSL get.docker.com -o get-docker.sh && sudo sh get-docker.sh --mirror Aliyun \
@@ -44,9 +46,11 @@ sudo cp ./configs/daemon.json /etc/docker \
 DOTFILES_DIR="$(pwd)/dotfiles"
 TARGET_DIR="$HOME"
 for file in "$DOTFILES_DIR"/.*; do
-  filename=$(basename "$file")
-  target_file="$TARGET_DIR/$filename"
-  ln -fs $file $target_file
+  if [ -f "$file" ]; then
+    filename=$(basename "$file")
+    target_file="$TARGET_DIR/$filename"
+    ln -fs $file $target_file
+  fi
 done
 
 # wsl
